move parseMultisigScript to bitcoin.util.ts
