+++ frontend/src/app/bitcoin.utils.ts
@@ -1,5 +1,4 @@
 import { Transaction, Vin } from './interfaces/electrs.interface';
-import { parseMultisigScript } from './components/address-labels/address-labels.component';
 
 const P2SH_P2WPKH_COST = 21 * 4; // the WU cost for the non-witness part of P2SH-P2WPKH
 const P2SH_P2WSH_COST  = 35 * 4; // the WU cost for the non-witness part of P2SH-P2WSH
@@ -112,6 +111,45 @@ export function calcSegwitFeeGains(tx: Transaction) {
     potentialTaprootGains: potentialTaprootGains / tx.weight,
     realizedTaprootGains: realizedTaprootGains / tx.weight
   };
+}
+
+/** extracts m and n from a multisig script (asm), returns nothing if it is not a multisig script */
+export function parseMultisigScript(script: string): void | { m: number, n: number } {
+  if (!script) {
+    return;
+  }
+  const ops = script.split(' ');
+  if (ops.length < 3 || ops.pop() !== 'OP_CHECKMULTISIG') {
+    return;
+  }
+  const opN = ops.pop();
+  if (!opN.startsWith('OP_PUSHNUM_')) {
+    return;
+  }
+  const n = parseInt(opN.match(/[0-9]+/)[0], 10);
+  if (ops.length < n * 2 + 1) {
+    return;
+  }
+  // pop n public keys
+  for (let i = 0; i < n; i++) {
+    if (!/^0((2|3)\w{64}|4\w{128})$/.test(ops.pop())) {
+      return;
+    }
+    if (!/^OP_PUSHBYTES_(33|65)$/.test(ops.pop())) {
+      return;
+    }
+  }
+  const opM = ops.pop();
+  if (!opM.startsWith('OP_PUSHNUM_')) {
+    return;
+  }
+  const m = parseInt(opM.match(/[0-9]+/)[0], 10);
+
+  if (ops.length) {
+    return;
+  }
+
+  return { m, n };
 }
 
 // https://github.com/shesek/move-decimal-point
