+++ frontend/src/app/components/address-labels/address-labels.component.ts
@@ -1,6 +1,7 @@
 import { Component, ChangeDetectionStrategy, Input, OnChanges } from '@angular/core';
 import { Vin, Vout } from '../../interfaces/electrs.interface';
 import { StateService } from 'src/app/services/state.service';
+import { parseMultisigScript } from 'src/app/bitcoin.utils';
 
 @Component({
   selector: 'app-address-labels',
@@ -108,42 +109,4 @@ export class AddressLabelsComponent implements OnChanges {
   handleVout() {
     this.detectMultisig(this.vout.scriptpubkey_asm);
   }
-}
-
-export function parseMultisigScript(script: string): void | { m: number, n: number } {
-  if (!script) {
-    return;
-  }
-  const ops = script.split(' ');
-  if (ops.length < 3 || ops.pop() !== 'OP_CHECKMULTISIG') {
-    return;
-  }
-  const opN = ops.pop();
-  if (!opN.startsWith('OP_PUSHNUM_')) {
-    return;
-  }
-  const n = parseInt(opN.match(/[0-9]+/)[0], 10);
-  if (ops.length < n * 2 + 1) {
-    return;
-  }
-  // pop n public keys
-  for (let i = 0; i < n; i++) {
-    if (!/^0((2|3)\w{64}|4\w{128})$/.test(ops.pop())) {
-      return;
-    }
-    if (!/^OP_PUSHBYTES_(33|65)$/.test(ops.pop())) {
-      return;
-    }
-  }
-  const opM = ops.pop();
-  if (!opM.startsWith('OP_PUSHNUM_')) {
-    return;
-  }
-  const m = parseInt(opM.match(/[0-9]+/)[0], 10);
-
-  if (ops.length) {
-    return;
-  }
-
-  return { m, n };
 }
